Basic code framework
Currently, the code will generate the Aruco markers, that can be printed out as well as play the input video, named "InputVideos/Input_2.mp4". Note that the video is not currently added in the repository. The video will be displayed, as will be the Aruco markers (well enough) visible in the video itself, frame by frame. The rest of the code has yet to be fully implemented.