Implement calibration
A limit was made for the number of images that will be taken into acount when computing the calibration matrix and distance coeficients.
The calibration data is written to a file, and if that file exists before the calibration process, the process will be skipped. This will greatly improve runtime between different runs of the code.

Note, the application currently stops and returns 42 when the calibration file exists.